requirements.txt
streamlit
qrcode
pillow
opencv-python
pyzbar
bcrypt
